# gateway-core/src/api/assets/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Label 309 Gateway API",
    "version": "1.0.0",
    "description": "The Label 309 Proof-of-Existence gateway HTTP API: quote, publish (single and batch), upload content for storage, list and retrieve on-chain records (everything a standalone verifier needs), read the account balance, and stream record/balance events.\n\nMost mutating endpoints require a Bearer API key carrying the relevant scope; the record collection and single-record reads are callable anonymously. The bearer key prefix, the error documentation base URL (the problem `type` member), and the server URLs are operator-configured per deployment.",
    "license": {
      "name": "Apache-2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "This deployment's data-plane base (the origin is operator-configured; the version segment is fixed)."
    }
  ],
  "paths": {
...
    "/poe/uploads/attempts/{attempt_id}": {
      "get": {
        "operationId": "poe.uploads.attempt",
        "tags": ["PoE"],
        "summary": "Poll the terminal outcome of an upload attempt.",
        "description": "The authoritative read of an upload attempt's outcome, for a client that received an `accepted` per-file result (its bytes were in flight under a concurrent upload) or whose original upload connection dropped. Returns `reserved` (still in flight — re-poll), `committed` (success, with the `uri` and `charged_usd_micros`), or `released` (failure, with a `reason`). The row is read after the terminal compare-and-set, so the result is never a dangling pointer. An attempt the caller does not own returns 404 indistinguishably from a non-existent one. This poll is AUTHORITATIVE for the outcome; the account event stream (GET /api/v1/account/balance/events) is a latency optimisation that merely prompts an earlier re-poll. Scope required: poe:create.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "The upload attempt id returned in an `accepted` per-file result."
          }
        ],
        "responses": {
          "200": {
            "description": "200 — the attempt's current state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PoeUploadAttempt"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundResponse"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitedResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalErrorResponse"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailableResponse"
          }
        }
      }
    },
...
    "/poe/publish": {
      "post": {
        "operationId": "poe.publish",
        "tags": ["PoE"],
        "summary": "Publish a finalised Label 309 record to Cardano. JSON only; storage URIs must already be substituted.",
        "description": "Accepts a finalised canonical-CBOR record (hex) plus the `quote_id` from a prior POST /api/v1/poe/quote, and submits the record to Cardano under metadata label 309. The record MUST already carry its final `ar://` (or `ipfs://`, when supported) URIs: files are uploaded first via POST /api/v1/poe/uploads and the returned URIs belong in the record before it is published; a structurally invalid record is rejected with 422 `validation-failed`. Dedups by `sha256(record_bytes)` per account so a retry of the exact same record returns the existing resource (200) without spawning a duplicate Cardano submission and without consuming the quote. New submissions return 202 with `id`, `status: \"submitting\"`, and `balance_after_usd_micros` (decimal string). The Cardano `tx_hash` is populated asynchronously by the cardano-confirm worker; subscribe to GET /api/v1/poe/events/{poe_id} for live updates. The 402 problem document embeds `balance_usd_micros` and `required_usd_micros` (decimal strings, micro-USD) as extension members so the client can show the payer exactly how short the account is. Scope required: poe:create.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255
            },
            "required": false,
            "name": "Idempotency-Key",
            "in": "header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PoePublishRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "200 — dedup hit; existing resource returned. No balance change.",
            "headers": {
              "RateLimit-Limit": {
                "$ref": "#/components/headers/RateLimit-Limit"
              },
              "RateLimit-Remaining": {
                "$ref": "#/components/headers/RateLimit-Remaining"
              },
              "RateLimit-Reset": {
                "$ref": "#/components/headers/RateLimit-Reset"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PoePublishResponse"
                }
              }
            }
          },
          "202": {
            "description": "202 — new record accepted; resource `id` assigned, tx_hash pending. Returns `balance_after_usd_micros` as a decimal string.",
            "headers": {
              "RateLimit-Limit": {
                "$ref": "#/components/headers/RateLimit-Limit"
              },
              "RateLimit-Remaining": {
                "$ref": "#/components/headers/RateLimit-Remaining"
              },
              "RateLimit-Reset": {
                "$ref": "#/components/headers/RateLimit-Reset"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PoePublishResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidBodyResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedResponse"
          },
          "402": {
            "$ref": "#/components/responses/InsufficientFundsResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "404": {
            "$ref": "#/components/responses/QuoteNotFoundResponse"
          },
          "409": {
            "$ref": "#/components/responses/QuoteAlreadyConsumedResponse"
          },
          "410": {
            "$ref": "#/components/responses/QuoteExpiredResponse"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailedResponse"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitedResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalErrorResponse"
          }
        }
      }
    },
...
    "/account/balance": {
      "get": {
        "operationId": "account.getBalance",
        "tags": ["Account"],
        "summary": "Read the caller's current USD balance as a one-shot JSON snapshot.",
        "description": "Returns the current prepaid USD balance for the caller's account as `{ \"balance_usd_micros\": \"<decimal string of micro-cents>\" }`. The value is a string (never a JSON number) so the bigint survives JSON without precision loss; an account with no ledger activity yet reads `\"0\"`. This is the one-shot complement to `GET /api/v1/account/balance/events` (which streams the same field live over SSE). Scope required: account:read.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "200 — current balance returned.",
            "headers": {
              "RateLimit-Limit": {
                "$ref": "#/components/headers/RateLimit-Limit"
              },
              "RateLimit-Remaining": {
                "$ref": "#/components/headers/RateLimit-Remaining"
              },
              "RateLimit-Reset": {
                "$ref": "#/components/headers/RateLimit-Reset"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountBalanceResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitedResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalErrorResponse"
          }
        }
      }
    },
    "/account/ledger": {
      "get": {
        "operationId": "account.listLedger",
        "tags": ["Account"],
        "summary": "List the caller's balance-ledger history, newest first.",
        "description": "The caller's own balance-ledger journal: every debit (publish, storage upload), credit (top-up adjustment, grant, refund), and reservation entry that moved the account's USD balance, newest first. Strictly account-scoped — a credential only ever sees its own account's entries. `amount_usd_micros` is a SIGNED decimal string (negative debit, positive credit; a string, never a JSON number, so the bigint survives JSON without precision loss). `kind` is the registered ledger kind (`poe_publish`, `storage_upload`, `storage_hold`, `storage_hold_release`, `refund_rollback`, `refund_user`, `manual_adjustment`, plus any vendor-registered kinds); `ref` is the entry's cross-reference key (the poe record id for a publish debit or refund, the upload attempt id for a storage charge, the vendor's idempotency key for an adjustment). Pagination uses an opaque base64url keyset cursor — callers must NOT crack it open. Scope required: account:read.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "cursor",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "200 — ledger page returned.",
            "headers": {
              "RateLimit-Limit": {
                "$ref": "#/components/headers/RateLimit-Limit"
              },
              "RateLimit-Remaining": {
                "$ref": "#/components/headers/RateLimit-Remaining"
              },
              "RateLimit-Reset": {
                "$ref": "#/components/headers/RateLimit-Reset"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountLedgerList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidCursorResponse"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenResponse"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitedResponse"
          },
          "500": {
            "$ref": "#/components/responses/InternalErrorResponse"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailableResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "API key authentication. Pass the operator-issued API key in the Authorization: Bearer header. The key's human-readable prefix and which deployment (network) it is scoped to are operator-configured."
      }
    },
    "schemas": {
      "PoePublishResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^poe_[0-9a-hjkmnp-tv-z]{26}$"
          },
          "tx_hash": {
            "type": ["string", "null"],
            "pattern": "^[0-9a-f]{64}$"
          },
          "status": {
            "type": "string",
            "enum": ["submitting", "submitted", "confirmed", "permanent_failure"]
          },
          "items_count": {
            "type": "integer",
            "minimum": 0
          },
          "signed": {
            "type": "boolean"
          },
          "sealed": {
            "type": "boolean"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "item_idx": {
                  "type": "integer",
                  "minimum": 0
                },
                "hashes": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string",
                    "pattern": "^[0-9a-f]+$"
                  }
                },
                "uris": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "enc": {
                  "type": "object",
                  "properties": {
                    "scheme": {
                      "type": "integer"
                    },
                    "aead": {
                      "type": "string"
                    },
                    "kem": {
                      "type": "string"
                    },
                    "nonce": {
                      "type": "string",
                      "pattern": "^[0-9a-f]+$"
                    },
                    "slots_count": {
                      "type": "integer",
                      "minimum": 0
                    },
                    "slots_mac": {
                      "type": "string",
                      "pattern": "^[0-9a-f]+$"
                    },
                    "passphrase": {
                      "type": "object",
                      "properties": {
                        "alg": {
                          "type": "string"
                        },
                        "salt": {
                          "type": "string",
                          "pattern": "^[0-9a-f]+$"
                        },
                        "params": {
                          "type": "object",
                          "additionalProperties": {}
                        }
                      },
                      "required": ["alg", "salt", "params"]
                    }
                  },
                  "additionalProperties": false
                }
              },
              "required": ["item_idx", "hashes"],
              "additionalProperties": false
            }
          },
          "conformance_profile": {
            "type": "string",
            "enum": ["core", "signed", "sealed", "recipient-sealed"]
          },
          "balance_after_usd_micros": {
            "type": "string",
            "pattern": "^[0-9]+$"
          }
        },
        "required": [
          "id",
          "tx_hash",
          "status",
          "items_count",
          "signed",
          "sealed",
          "items",
          "conformance_profile",
          "balance_after_usd_micros"
        ],
        "additionalProperties": false
      },
...
      "PoeUploadAttempt": {
        "type": "object",
        "description": "The current state of an upload attempt, the authoritative terminal-outcome read. `reserved` is in flight; `committed` is the terminal success (carries `uri` and `charged_usd_micros`); `released` is the terminal failure (carries a `reason`).",
        "properties": {
          "attempt_id": {
            "type": "string",
            "format": "uuid"
          },
          "state": {
            "type": "string",
            "enum": ["reserved", "committed", "released"]
          },
          "sha256": {
            "type": "string",
            "pattern": "^[0-9a-f]+$"
          },
          "bytes": {
            "type": "integer",
            "minimum": 0
          },
          "backend": {
            "type": "string"
          },
          "uri": {
            "type": "string",
            "description": "The addressable content URI. Present only on a `committed` attempt."
          },
          "charged_usd_micros": {
            "type": "integer",
            "minimum": 0,
            "description": "The storage USD (micro-USD) ACTUALLY debited for this attempt, not the reserve-time estimate: 0 when the upload committed but deduped against bytes already stored for this account on this backend. Present only on a `committed` attempt."
          },
          "reason": {
            "type": "string",
            "enum": ["provider_rejected", "unrecoverable_staged_content_lost"],
            "description": "Why the attempt failed. Present only on a `released` attempt: `provider_rejected` (the provider refused the upload; retry) or `unrecoverable_staged_content_lost` (the bytes did not survive a crash and the provider never received them; re-upload the content)."
          }
        },
        "required": ["attempt_id", "state", "sha256", "bytes", "backend"],
        "additionalProperties": false
      },
...
      "PoePublishRequest": {
        "type": "object",
        "properties": {
          "record": {
            "type": "string",
            "pattern": "^[0-9a-f]+$"
          },
          "quote_id": {
            "type": "string",
            "format": "uuid"
          },
          "signatures": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "cose_sign1": {
                  "type": "string",
                  "pattern": "^[0-9a-f]+$"
                },
                "cose_key": {
                  "type": "string",
                  "pattern": "^[0-9a-f]+$"
                }
              },
              "required": ["cose_sign1"],
              "additionalProperties": false
            }
          }
        },
        "required": ["record", "quote_id"],
        "additionalProperties": false
      },
...
    },
    "responses": {
...
      "ServiceUnavailableResponse": {
        "description": "503 — service-unavailable",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            },
            "example": {
              "type": "https://gateway.example/api/v1/errors#service-unavailable",
              "title": "Service unavailable",
              "status": 503,
              "detail": "A dependency the request needs (storage, the database) is temporarily unavailable. Retry after a short backoff.",
              "code": "service-unavailable",
              "trace_id": "01977c8a-7c5e-7000-8000-000000000000"
            }
          }
        }
      },
      "ForbiddenResponse": {
        "description": "403 — insufficient-scope. The problem body carries the RFC 7807 §3.2 extension members `required` (string[]: the scope the endpoint requires) and `granted` (string[]: the scopes the credential actually carries — empty for an operator credential presented to the data plane, which grants nothing there).",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            },
            "example": {
              "type": "https://gateway.example/api/v1/errors#insufficient-scope",
              "title": "Insufficient Scope",
              "status": 403,
              "detail": "this endpoint requires the poe:create scope",
              "code": "insufficient-scope",
              "required": ["poe:create"],
              "granted": ["poe:read"],
              "trace_id": "01977c8a-7c5e-7000-8000-000000000000"
            }
          }
        }
      },
      "UnauthorizedResponse": {
        "description": "401 — unauthorized",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            },
            "example": {
              "type": "https://gateway.example/api/v1/errors#unauthorized",
              "title": "Unauthorized",
              "status": 401,
              "detail": "Authentication is required to access this resource. The `code` field is always `unauthorized`: a missing Authorization header, a malformed Bearer credential, and an unknown or revoked one all collapse to the same response, so a scanner cannot distinguish credential formats from live secrets.",
              "code": "unauthorized",
              "trace_id": "01977c8a-7c5e-7000-8000-000000000000"
            }
          }
        }
      },
...
      "InternalErrorResponse": {
        "description": "500 — internal-error",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            },
            "example": {
              "type": "https://gateway.example/api/v1/errors#internal-error",
              "title": "Internal Error",
              "status": 500,
              "detail": "An unexpected server error occurred.",
              "code": "internal-error",
              "trace_id": "01977c8a-7c5e-7000-8000-000000000000"
            }
          }
        }
      },
...
      "NotFoundResponse": {
        "description": "404 — not-found",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            },
            "example": {
              "type": "https://gateway.example/api/v1/errors#not-found",
              "title": "Not Found",
              "status": 404,
              "detail": "No PoE record is indexed under that transaction hash.",
              "code": "not-found",
              "trace_id": "01977c8a-7c5e-7000-8000-000000000000"
            }
          }
        }
      },
...
      "RateLimitedResponse": {
        "description": "429 — rate-limited",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetails"
            },
            "example": {
              "type": "https://gateway.example/api/v1/errors#rate-limited",
              "title": "Rate Limited",
              "status": 429,
              "detail": "API-key rate limit exceeded.",
              "code": "rate-limited",
              "trace_id": "01977c8a-7c5e-7000-8000-000000000000"
            }
          }
        }
      },
...
    },
    "headers": {
      "RateLimit-Reset": {
        "description": "IETF draft-ietf-httpapi-ratelimit-headers-09. Seconds until the current sliding window rolls over — the earliest moment `RateLimit-Remaining` could fully reset.",
        "schema": {
          "type": "integer",
          "minimum": 1
        }
      },
      "RateLimit-Remaining": {
        "description": "IETF draft-ietf-httpapi-ratelimit-headers-09. Tokens left in the current sliding window after this request was counted. Zero on a 429 response.",
        "schema": {
          "type": "integer",
          "minimum": 0
        }
      },
      "RateLimit-Limit": {
        "description": "IETF draft-ietf-httpapi-ratelimit-headers-09. The per-window request ceiling enforced for the API key on this endpoint, in requests per 60 seconds.",
        "schema": {
          "type": "integer",
          "minimum": 1
        }
      }
    }
  }
}